third_party/WebKit/LayoutTests/external/wpt/fetch/api/request/request-cache-no-store-expected.txt

This is a testharness.js-based test.
FAIL RequestCache "no-store" mode does not check the cache for previously cached content and goes to the network regardless with Etag and stale response assert_equals: expected (undefined) undefined but got (string) "\"0.7131790571494467\""
FAIL RequestCache "no-store" mode does not check the cache for previously cached content and goes to the network regardless with Last-Modified and stale response assert_equals: expected (undefined) undefined but got (string) "Mon, 27 Mar 2017 11:24:51 GMT"
FAIL RequestCache "no-store" mode does not check the cache for previously cached content and goes to the network regardless with Etag and fresh response assert_equals: expected 2 but got 1
FAIL RequestCache "no-store" mode does not check the cache for previously cached content and goes to the network regardless with Last-Modified and fresh response assert_equals: expected 2 but got 1
FAIL RequestCache "no-store" mode does not store the response in the cache with Etag and stale response assert_equals: expected (undefined) undefined but got (string) "\"0.47496026068576236\""
FAIL RequestCache "no-store" mode does not store the response in the cache with Last-Modified and stale response assert_equals: expected (undefined) undefined but got (string) "Mon, 27 Mar 2017 11:24:51 GMT"
FAIL RequestCache "no-store" mode does not store the response in the cache with Etag and fresh response assert_equals: expected 2 but got 1
FAIL RequestCache "no-store" mode does not store the response in the cache with Last-Modified and fresh response assert_equals: expected 2 but got 1
Harness: the test ran to completion.
